Tpetra: FEM Assembly to common main function
Consolidating FEM Assembly examples into a single main with
selection of algorithm via command line parameters.

Relates to issue #2369
