+++ packages/tpetra/core/example/Finite-Element-Assembly/fem_assembly_utility.hpp
@@ -78,5 +78,5 @@ getOutputStream(const Teuchos::Comm<int>& comm)
 
 } // namespace TpetraExamples
 
-#endif
+#endif  // FEM_ASSEMBLY_UTILITY_HPP
